+++ CHANGELOG.md
@@ -63,7 +63,7 @@
 
 - Update to [`libp2p-kad` `v0.41.0`](protocols/kad/CHANGELOG.md#0410).
 - 
-- Update to [`libp2p-identify` `v0.40.0`](protocols/identify/CHANGELOG.md#0400).
+- Update to [`libp2p-identify` `v0.39.1`](protocols/identify/CHANGELOG.md#0400).
 
 - Update to [`libp2p-noise` `v0.39.1`](transports/noise/CHANGELOG.md#0391).