+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - [`libp2p-uds` CHANGELOG](transports/uds/CHANGELOG.md)
 - [`libp2p-wasm-ext` CHANGELOG](transports/wasm-ext/CHANGELOG.md)
 - [`libp2p-websocket` CHANGELOG](transports/websocket/CHANGELOG.md)
+- [`libp2p-tls` CHANGELOG](transports/tls/CHANGELOG.md)
 
 ## Multiplexers
 
@@ -68,6 +69,8 @@
 - Update to [`libp2p-noise` `v0.39.1`](transports/noise/CHANGELOG.md#0391).
 
 - Update to [`libp2p-swarm` `v0.39.1`](swarm/CHANGELOG.md#0391).
+
+- Introduce [`libp2p-tls` `v0.1.0`](transports/tls/CHANGELOG.md#010).
 
 # 0.48.0