+++ CHANGELOG.md
@@ -57,6 +57,8 @@
 - Update to [`libp2p-metrics` `v0.10.0`](misc/metrics/CHANGELOG.md#0100).
 
 - Update to [`libp2p-kad` `v0.41.0`](protocols/kad/CHANGELOG.md#0410).
+- 
+- Update to [`libp2p-identify` `v0.39.1`](protocols/identify/CHANGELOG.md#0400).
 
 - Update to [`libp2p-autonat` `v0.8.0`](protocols/autonat/CHANGELOG.md#0080).
 
@@ -99,6 +101,8 @@
 - Update to [`libp2p-gossipsub` `v0.42.0`](protocols/gossipsub/CHANGELOG.md#0420).
 
 [PR 2918]: https://github.com/libp2p/rust-libp2p/pull/2918
+
+- Update to [`libp2p-swarm` `v0.39.1`](swarm/CHANGELOG.md#0391).
 
 # 0.48.0