+++ CHANGELOG.md
@@ -77,6 +77,8 @@
 
 - [PR 2918]: https://github.com/libp2p/rust-libp2p/pull/2918
 
+- Update to [`libp2p-identify` `v0.39.1`](protocols/identify/CHANGELOG.md#0391).
+
 # 0.48.0
 
 - Update to [`libp2p-core` `v0.36.0`](core/CHANGELOG.md#0360).